Move risk histogram code into common location for reuse
