+++ simulatingrisk/charts/histogram.js
