--- simulatingrisk/risky_bet/HistogramModule.js
@@ -1,1 +1,0 @@
-../risky_food/HistogramModule.js
